Add a +1 when commenting /validate
This commits extracts the verification that the content of the comment is /validate so that the reaction is added as soon as possible.
Here, it is the first thing that is done then the event type is issue_comment. If the reaction was a step of the parse-issue job, the wait would have been longer.